% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssistLeaderByYear.R
\name{getAssistLeaderByYear}
\alias{getAssistLeaderByYear}
\title{getAssistLeaderByYear}
\usage{
getAssistLeaderByYear(year)
}
\arguments{
\item{year}{}
}
\description{
gets assist leader by year
}
\examples{

}
